+++ src/common/Environment.ts
@@ -66,14 +66,14 @@ args
     ["", "disableProductStyleUrl"],
     "Optional. Disable getting account name from the host of request Uri, always get account name from the first path segment of request Uri."
   )
-  .option(
-    ["", "inMemoryPersistence"],
-    "Optional. Disable persisting any data to disk. If the Azurite process is terminated, all data is lost."
-  )
   .option(["", "oauth"], 'Optional. OAuth level. Candidate values: "basic"')
   .option(["", "cert"], "Optional. Path to certificate file")
   .option(["", "key"], "Optional. Path to certificate key .pem file")
   .option(["", "pwd"], "Optional. Password for .pfx file")
+  .option(
+    ["", "inMemoryPersistence"],
+    "Optional. Disable persisting any data to disk. If the Azurite process is terminated, all data is lost."
+  )
   .option(
     ["d", "debug"],
     "Optional. Enable debug log by providing a valid local file path as log destination"