+++ src/queue/QueueEnvironment.ts
@@ -40,6 +40,10 @@ args
   .option(["", "key"], "Optional. Path to certificate key .pem file")
   .option(["", "pwd"], "Optional. Password for .pfx file")
   .option(
+    ["", "inMemoryPersistence"],
+    "Optional. Disable persisting any data to disk. If the Azurite process is terminated, all data is lost."
+  )
+  .option(
     ["d", "debug"],
     "Optional. Enable debug log by providing a valid local file path as log destination"
   );