+++ src/table/TableEnvironment.ts
@@ -39,6 +39,10 @@ args
     "Optional. Disable getting account name from the host of request Uri, always get account name from the first path segment of request Uri."
   )
   .option(
+    ["", "inMemoryPersistence"],
+    "Optional. Disable persisting any data to disk. If the Azurite process is terminated, all data is lost."
+  )
+  .option(
     ["d", "debug"],
     "Optional. Enable debug log by providing a valid local file path as log destination"
   )